ON-981 - Now using POST & ViewModels. Also started pre-search journey.

+++ src/SFA.DAS.ApplyService.InternalApi/Models/AssessorService/OrganisationSummary.cs
@@ -4,7 +4,7 @@
     {
         public string Id { get; set; }
         public string Name { get; set; }
-        public long? Ukprn { get; set; }
+        public int? Ukprn { get; set; }
 
         public OrganisationData OrganisationData { get; set; }
